+++ beetroot/exception.py
@@ -37,3 +37,6 @@ class StupidProgrammerError(Exception):
 
 class StringError(Exception):
     pass
+
+class InvalidFileError(Exception):
+    pass