+++ CHANGELOG.rst
@@ -89,6 +89,7 @@ Version number \= 1.Major.Minor.Patch
   Used kwargs in file.jdump()
   Merged file.bdump() with file.dump()
   Added annotations
-  Added lambda math functions
-  Added math class
   Updated crash() to use bugs instead of taskkill()
+  
+- 1.1.6.1 ; Added math class
+  Updated crash() to work properly